Power Endpoints now Follow Slimmed Controller Pattern
