+++ api/ExpressedRealms.Knowledges.API/GetKnowledgeTypes/KnowledgeTypeResponse.cs
@@ -1,4 +1,4 @@
-using ExpressedRealms.Powers.API.PowerEndpoints.Responses.Options;
+using ExpressedRealms.Powers.API.PowerEndpoints.GetPowerOptions;
 
 namespace ExpressedRealms.Knowledges.API.GetKnowledgeTypes;
