+++ api/ExpressedRealms.Powers.API/PowerEndpoints/GetPower/EditPowerInformationResponse.cs
@@ -1,6 +1,6 @@
-namespace ExpressedRealms.Powers.API.PowerEndpoints.Requests.PowerEdit;
+namespace ExpressedRealms.Powers.API.PowerEndpoints.GetPower;
 
-public class EditPowerRequest
+public class EditPowerInformationResponse
 {
     public int Id { get; set; }
     public required string Name { get; set; }
@@ -8,11 +8,12 @@ public class EditPowerRequest
     public required string Description { get; set; }
     public required string GameMechanicEffect { get; set; }
     public string? Limitation { get; set; }
-    public byte PowerDurationId { get; set; }
-    public byte AreaOfEffectId { get; set; }
+    public int PowerDurationId { get; set; }
+    public int AreaOfEffectId { get; set; }
     public int PowerLevelId { get; set; }
-    public byte PowerActivationTypeId { get; set; }
+    public int PowerActivationTypeId { get; set; }
     public string? Other { get; set; }
     public bool IsPowerUse { get; set; }
     public string? Cost { get; set; }
+    public int? StatModifierGroup { get; set; }
 }
