+++ api/ExpressedRealms.Powers.API/PowerEndpoints/GetPowersForPowerPath/DetailedInformation.cs
@@ -1,4 +1,4 @@
-namespace ExpressedRealms.Powers.API.PowerEndpoints.Responses.PowerList;
+namespace ExpressedRealms.Powers.API.PowerEndpoints.GetPowersForPowerPath;
 
 public class DetailedInformation
 {
