+++ api/ExpressedRealms.Powers.API/PowerEndpoints/GetPowersForPowerPath/PrerequisiteDisplay.cs
@@ -1,4 +1,4 @@
-namespace ExpressedRealms.Powers.API.PowerEndpoints.Responses.PowerList;
+namespace ExpressedRealms.Powers.API.PowerEndpoints.GetPowersForPowerPath;
 
 public class PrerequisiteDisplay
 {
